requirements.txt
asgiref==3.5.0
dj-database-url==1.0.0
Django==4.0.1
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
psycopg2-binary==2.9.5
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.2.0
